Handle temporary download paths in a better way
Rather than downloading to randomly generated filename and trying to append a suffix, try to use the URL as a filename. This should effectively be unique.

Day One chokes on attachment filenames without suffixes, so if it doesn't seem to have one, make one up.

There's a theoretical race condition if two posts have the same attachment URL and are being processed at the same time, but currently it's sequential so that's impossible.